+++ src/QueryBuilder/Condition/Builder/LikeBuilder.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\QueryBuilder\Condition\Builder;
 
+use Traversable;
 use Yiisoft\Db\Command\Param;
 use Yiisoft\Db\Constant\DataType;
 use Yiisoft\Db\Exception\Exception;
@@ -15,27 +16,25 @@ use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\QueryBuilder\Condition\Like;
 use Yiisoft\Db\QueryBuilder\Condition\LikeConjunction;
 use Yiisoft\Db\QueryBuilder\Condition\LikeMode;
+use Yiisoft\Db\QueryBuilder\Condition\NotLike;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 use function implode;
-use function is_array;
-use function preg_match;
+use function is_string;
 use function str_contains;
-use function strtoupper;
 use function strtr;
 
 /**
- * Build an object of {@see Like} into SQL expressions.
+ * Build an object of {@see Like} or {@see NotLike} into SQL expressions.
  *
- * @implements ExpressionBuilderInterface<Like>
+ * @implements ExpressionBuilderInterface<Like|NotLike>
  */
 class LikeBuilder implements ExpressionBuilderInterface
 {
-    public function __construct(
-        private readonly QueryBuilderInterface $queryBuilder,
-        private readonly string|null $escapeSql = null
-    ) {
-    }
+    /**
+     * @var string SQL fragment to append to the end of `LIKE` conditions.
+     */
+    protected const ESCAPE_SQL = '';
 
     /**
      * @var array Map of chars to their replacements in `LIKE` conditions. By default, it's configured to escape
@@ -47,10 +46,15 @@ class LikeBuilder implements ExpressionBuilderInterface
         '\\' => '\\\\',
     ];
 
+    public function __construct(
+        private readonly QueryBuilderInterface $queryBuilder,
+    ) {
+    }
+
     /**
-     * Build SQL for {@see Like}.
+     * Build SQL for {@see Like} or {@see NotLike}.
      *
-     * @param Like $expression
+     * @param Like|NotLike $expression
      *
      * @throws Exception
      * @throws InvalidArgumentException
@@ -61,24 +65,30 @@ class LikeBuilder implements ExpressionBuilderInterface
     {
         $values = $expression->value;
 
-        [$not, $operator] = $this->parseOperator($expression);
+        [$not, $operator] = $this->getOperatorData($expression);
 
-        if (!is_array($values)) {
-            $values = [$values];
+        if ($values === null) {
+            return $this->buildForEmptyValue($not);
         }
 
-        if (empty($values)) {
-            return $not ? '' : '0=1';
+        if (is_iterable($values)) {
+            if ($values instanceof Traversable) {
+                $values = iterator_to_array($values);
+            }
+            if (empty($values)) {
+                return $this->buildForEmptyValue($not);
+            }
+        } else {
+            $values = [$values];
         }
 
         $column = $this->prepareColumn($expression, $params);
 
         $parts = [];
-
-        /** @psalm-var list<string|ExpressionInterface> $values */
         foreach ($values as $value) {
+            /** @var ExpressionInterface|int|string $value */
             $placeholderName = $this->preparePlaceholderName($value, $expression, $params);
-            $parts[] = "$column $operator $placeholderName$this->escapeSql";
+            $parts[] = "$column $operator $placeholderName" . static::ESCAPE_SQL;
         }
 
         $conjunction = match ($expression->conjunction) {
@@ -97,7 +107,7 @@ class LikeBuilder implements ExpressionBuilderInterface
      * @throws InvalidConfigException
      * @throws NotSupportedException
      */
-    protected function prepareColumn(Like $condition, array &$params): string
+    protected function prepareColumn(Like|NotLike $condition, array &$params): string
     {
         $column = $condition->column;
 
@@ -122,15 +132,15 @@ class LikeBuilder implements ExpressionBuilderInterface
      * @return string
      */
     protected function preparePlaceholderName(
-        string|ExpressionInterface $value,
-        Like $condition,
+        string|int|ExpressionInterface $value,
+        Like|NotLike $condition,
         array &$params,
     ): string {
         if ($value instanceof ExpressionInterface) {
             return $this->queryBuilder->buildExpression($value, $params);
         }
 
-        if ($condition->escape) {
+        if (is_string($value) && $condition->escape) {
             $value = strtr($value, $this->escapingReplacements);
         }
 
@@ -138,29 +148,27 @@ class LikeBuilder implements ExpressionBuilderInterface
             LikeMode::Contains => '%' . $value . '%',
             LikeMode::StartsWith => $value . '%',
             LikeMode::EndsWith => '%' . $value,
-            LikeMode::Custom => $value,
+            LikeMode::Custom => (string) $value,
         };
 
         return $this->queryBuilder->bindParam(new Param($value, DataType::STRING), $params);
     }
 
     /**
-     * Parses operator and returns its parts.
-     *
-     * @throws InvalidArgumentException
+     * Get operator and `not` flag for the given condition.
      *
      * @psalm-return array{0: bool, 1: string}
      */
-    protected function parseOperator(Like $condition): array
+    protected function getOperatorData(Like|NotLike $condition): array
     {
-        $operator = strtoupper($condition->operator);
-        if (!preg_match('/^((NOT |)I?LIKE)/', $operator, $matches)) {
-            throw new InvalidArgumentException("Invalid operator in like condition: \"$operator\"");
-        }
+        return match ($condition::class) {
+            Like::class => [false, 'LIKE'],
+            NotLike::class => [true, 'NOT LIKE'],
+        };
+    }
 
-        $not = !empty($matches[2]);
-        $operator = $matches[1];
-
-        return [$not, $operator];
+    private function buildForEmptyValue(bool $not): string
+    {
+        return $not ? '' : '0=1';
     }
 }